chore: post create, update and utils:

+++ api/controller/post.controller.js
@@ -1,29 +1,32 @@
 import prisma from '../lib/prisma.js';
-import { successRequest } from '../utils/request-utils.js';
+import { errorHandler, successRequest } from '../utils/request-utils.js';
 
-export const createPost = async (req, res) => {
+export const createPost = async (req, res, next) => {
     try {
-        const post = await prisma.post.create({ ...req.body });
+        const post = await prisma.post.create({
+            data: {
+                ...req.body,
+                userId: req.userId
+            }
+        });
 
         return successRequest(res, 200, post)
     } catch (e) {
-        console.log(e)
-        next("Something went wrong", 500)
+        next(e)
     }
 }
 
-export const getPosts = async (req, res) => {
+export const getPosts = async (req, res, next) => {
     try {
         const posts = await prisma.post.findMany({});
 
         return successRequest(res, 200, posts)
     } catch (e) {
-        console.log(e)
-        next("Something went wrong", 500)
+        next(e)
     }
 }
 
-export const getPostById = async (req, res) => {
+export const getPostById = async (req, res, next) => {
     try {
         const post = await prisma.post.findUnique({
             where: {
@@ -31,14 +34,18 @@ export const getPostById = async (req, res) => {
             }
         });
 
+        if (!post) {
+            throw new Error("Cannot find the post")
+        }
+
         return successRequest(res, 200, post)
     } catch (e) {
-        console.log(e)
-        next("Something went wrong", 500)
+        console.log("error", e)
+        next(e)
     }
 }
 
-export const updatePostById = async (req, res) => {
+export const updatePostById = async (req, res, next) => {
     try {
         const post = await prisma.post.findUnique({
             where: {
@@ -49,6 +56,10 @@ export const updatePostById = async (req, res) => {
 
         if (!post) {
             next("Post not found", 404)
+        }
+
+        if (post.userId !== req.userId) {
+            next("You are not authorized", 403)
         }
 
         const updatedPost = await prisma.post.update({
@@ -62,13 +73,12 @@ export const updatePostById = async (req, res) => {
 
         return successRequest(res, 200, updatedPost)
     } catch (e) {
-        console.log(e)
-        next("Something went wrong", 500)
+        next(e)
     }
 }
 
 
-export const deletePostById = async (req, res) => {
+export const deletePostById = async (req, res, next) => {
     try {
         const post = await prisma.post.findUnique({
             where: {
@@ -76,18 +86,22 @@ export const deletePostById = async (req, res) => {
             }
         });
 
-
         if (!post) {
             next("Post not found", 404)
         }
 
+        if (post.userId !== req.userId) {
+            next("You are not authorized", 403)
+        }
+
         const deletedPost = await prisma.post.delete({
-            where: req.params.id
+            where: {
+                id: req.params.id
+            }
         })
 
         return successRequest(res, 201, deletedPost)
     } catch (e) {
-        console.log(e)
-        next("Something went wrong", 500)
+        next(e)
     }
 }
